feat: adicionados erros e testados

+++ src/modules/users/users.service.ts
@@ -2,6 +2,7 @@ import {
   Injectable,
   ConflictException,
   UnauthorizedException,
+  NotFoundException,
 } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
@@ -27,20 +28,30 @@ export class UsersService {
     return this.usersRepository.findAll();
   }
 
-  async findOne(userId: string) {
-    return await this.usersRepository.findOne(parseInt(userId));
+  async findOne(userId: number) {
+    return await this.usersRepository.findOne(userId);
   }
 
   async findByEmail(email: string) {
     return await this.usersRepository.findByEmail(email);
   }
 
-  update(id: string, updateUserDto: UpdateUserDto) {
-    return this.usersRepository.update(id, updateUserDto);
+  async update(userId: number, email: string, updateUserDto: UpdateUserDto) {
+    const findUser = await this.usersRepository.findOne(userId);
+
+    if (!findUser) {
+      throw new NotFoundException({ message: 'User not found' });
+    }
+    return await this.usersRepository.update(userId, updateUserDto);
   }
 
-  remove(id: string) {
-    return this.usersRepository.delete(id);
+  async remove(id: number) {
+    const findUser = await this.usersRepository.findOne(id);
+
+    if (!findUser) {
+      throw new NotFoundException({ message: 'User not found' });
+    }
+    return await this.usersRepository.delete(id);
   }
 
   async validateStaff(email: string) {
